Update code to be in AwardFormService(SubmissionFormService)
